Convert from icon font to SVG icons
Update to the CF dependencies that are now using SVG icons and replaces
the old class based pseudo element icons with inline icons via a custom
svg_icon tag and update the Less and JS to work with the new inline
icons. Uses the new mock staticfiles finder to allow the `svg_icon` tag
to render in tests where the icons don't exist.

Co-authored-by: Scott Cranfill <scott.cranfill@cfpb.gov>
Co-authored-by: Andy Chosak <andy@chosak.org>