Properly indent inline interp text
We use CSS classes to indent paragraph text. To assign the correct
class, we count the number of dashes in the label to determine
how deeply nested the paragraph is. Inline interps have special
prefixes that are removed before counting the dashes.
e.g. 6-a-Interp-1 becomes -1 and gets a `level-1` class
e.g. 12-b-Interp-2-i becomes -2-i and gets a `level-2` class